+++ library/src/main/java/org/mustangproject/ReferencedDocument.java
@@ -2,6 +2,8 @@ package org.mustangproject;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
+import java.util.Date;
+
 import org.mustangproject.ZUGFeRD.IReferencedDocument;
 import org.mustangproject.util.NodeMap;
 import org.w3c.dom.Node;
@@ -13,20 +15,33 @@ public class ReferencedDocument implements IReferencedDocument {
 	String issuerAssignedID;
 	String typeCode;
 	String referenceTypeCode;
+	Date formattedIssueDateTime;
 
 	public ReferencedDocument(String issuerAssignedID, String typeCode, String referenceTypeCode) {
-		this.issuerAssignedID = issuerAssignedID;
+		this(issuerAssignedID);
 		this.typeCode = typeCode;
 		this.referenceTypeCode = referenceTypeCode;
 	}
 
-	public ReferencedDocument(String issuerAssignedID, String referenceTypeCode) {
-		this.issuerAssignedID = issuerAssignedID;
-		this.typeCode = "916"; // additional invoice related document
-		this.referenceTypeCode = referenceTypeCode;
+	public ReferencedDocument(String issuerAssignedID, String typeCode, String referenceTypeCode, Date formattedIssueDateTime) {
+		this(issuerAssignedID, typeCode, referenceTypeCode);
+		this.formattedIssueDateTime = formattedIssueDateTime;
 	}
 
-    /***
+	public ReferencedDocument(String issuerAssignedID, String referenceTypeCode) {
+		this(issuerAssignedID, "916", referenceTypeCode); // additional invoice related document
+	}
+
+	public ReferencedDocument(String issuerAssingedID, Date formattedIssueDateTime) {
+		this(issuerAssingedID);
+		this.formattedIssueDateTime = formattedIssueDateTime;
+	}
+	
+	public ReferencedDocument(String issuerAssignedID) {
+		this.issuerAssignedID = issuerAssignedID;
+	}
+
+	/***
 	 * sets an ID assigned by the sender
 	 * @param issuerAssignedID the ID as a string :-)
 	 */
@@ -50,6 +65,15 @@ public class ReferencedDocument implements IReferencedDocument {
 		this.referenceTypeCode = referenceTypeCode;
 	}
 
+	/**
+	 * issue date of this line
+	 *
+	 * @param formattedIssueDateTime as Date
+	 */
+	public void setFormattedIssueDateTime(Date formattedIssueDateTime) {
+		this.formattedIssueDateTime = formattedIssueDateTime;
+	}
+
 	@Override
 	public String getIssuerAssignedID() {
 		return issuerAssignedID;
@@ -65,6 +89,12 @@ public class ReferencedDocument implements IReferencedDocument {
 		return referenceTypeCode;
 	}
 
+	@Override
+	public Date getFormattedIssueDateTime()
+	{
+		return formattedIssueDateTime;
+	}
+
 	public static ReferencedDocument fromNode(Node node) {
 		if (!node.hasChildNodes()) {
 			return null;
@@ -72,6 +102,7 @@ public class ReferencedDocument implements IReferencedDocument {
 		NodeMap nodes = new NodeMap(node);
 		return new ReferencedDocument(nodes.getAsStringOrNull("IssuerAssignedID", "ID"),
 			nodes.getAsStringOrNull("TypeCode", "DocumentTypeCode"),
-			nodes.getAsStringOrNull("ReferenceTypeCode"));
+			nodes.getAsStringOrNull("ReferenceTypeCode"),
+			XMLTools.tryDate(nodes.getAsStringOrNull("FormattedIssueDateTime")));
 	}
 }