+++ History.md
@@ -1,4 +1,25 @@
-- correcly write charge reason codes also for non-Xrechnung
+- 639
+- 633
+-626,
+-622,
+-356
+allow to add includedNotes with type
+- - 645
+- 631  multiple invoice referenced documents
+- 629
+- 630 #296 #565
+- 
+
+2.15.2
+=======
+2024-12-19
+- correcly write charge reason codes also for non-Xrechnung #617
+- correctly import additional referenced documents into invoice/corrected setting of attachments from jackson
+- corrected parseException structure
+- allow 1p0 as potential xmp version number
+- #618 import BT-20
+- #599 add tax category code for free export
+- #600 Fixes a problem where a stream was not safely closed
 
 2.15.1
 =======